All tool configurations are now handeld by the ToolSetupService
