+++ Assets/i5 Toolkit for Mixed Reality/Runtime/PieMenu/Scripts/ToolSetup.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: fdd752a8ba802614ea8ca814973b8a84
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
